fix: fix read stuck when reenter mountpoint
When avfsd reads files, it acquires a global lock. If a request causes
avfsd to re-enter the mount point and attempt to read, it will cause a
second locking event, resulting in a block. The solution is to prevent
the fuse daemon from reading itself.

Log: fix read stuck when reenter mountpoint
Bug: https://pms.uniontech.com/bug-view-333583.html